Multiple updates
Removed function to only do Nmap script scan on open ports only. It wasn't adding efficiency. 

Added a function that creates a file called "Versions" with the name and version number of the open port (if available). Helpful for quickly double checking later on, or to have something to feed into another service/vulnerability scanner.